requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
pandas
pathlib
scipy